Allow for granular update worker control
This commit allows for workers to be disabled at both
a global installation level, as well as on a per api
key basis. Calls added to the 'disabled_workers'
array in eveapi.config are disabled globally for all
updates. Calls added to the 'disabled_calls' field
in the EveApiKey model (json_encoded) will be ignored
for the affected key only.