Improve queue job handling.
This is done by adding a small 3 second delay to the time when a
queued job will be considered ready for release, allowing us to
write a tracking entry to the database. Previously, this was actually
a small race condition that never ended well at all.

A log entry will also be written now when a job_id returned from the
queued job dispatcher of the expected length.