feat(prices): improve price management
in order to provide more exact figures, we relay on a new endpoint to collect item prices.

market history endpoint is designed to provide daily stats per item into requested region.

we also introduce a new setting which allow instance owner to define the region from which prices have to be pulled from.