feat: bypass cache control in case we have no data in DB
in certain case, we might be in the following situation:
there is a cached entry related to a job call which is still valid (either the expiration date/time is not reached or ESI tell us cached entry is unchanged)

as a result, jobs will not process data and be mark as completed.

if that would be expectable, in case the related entry have been dropped (ie: deleted user, deleted character, deleted corporation, ...) - an admin will have to clear cache to allow those data to populate database again.