fix: update character on invalid member status
when ESI is returning an error telling a character is non longer member from the requested corporation, remove his affiliation and corporation member tracking entry from the system.

this is to avoid an invalid loop request targeting a corporation using the token from a character which is non longer part of a corporation - but that SeAT isn't notified yet.