fix(jobs): better handle esi errors
non longer throw an exception in the failed method, which will be used by workers when an exception is already throw.

doing this is unnecessary.

also, switch esi error handling to EsiBase since it's not a case valid for all jobs.

use listener to prevent excessive invalid calls against ESI.

when sso is doing strange things with tokens, we rollback the expiration date/time - so eseye will be able to renew the token on next call.

last but no least, we spawn a new eseye client at every new attempt - doing this will ensure us to always use up-to-date token.