fix(db): mitigate MariaDB issue with json field
This patch will issue a non-altering, alter table statement if the column addition fails. This statement will allow the migration to pass. It is unclear at this stage why this works.

for reference : https://jira.mariadb.org/browse/MDEV-25672

This change should be rollback when new MariaDB 10.5.11 will be released.