fix(jobs): Use Cache facade instead of Redis facade
The key name should not be hardcoded. Because a complete key name contains a prefix, and this prefix is not necessarily `seat`, it depends on the config [0].

As an impact, CheckEsiRateLimit middleware may not work. Since EsiBase::getRateLimitKeyTtl() always returns -2, EsiBase::incrementEsiRateLimit() will always set the cache to 1. Ultimately, CheckEsiRateLimit middleware will not do anything even if the error rate has actually been reached.

[0] Illuminate\Cache\CacheManager::getPrefix()