fix(jobs): address requests throttling issue
ensure history job is throttling ESI requests rather than itself.
chaining all market history jobs are chained all together instead of delaying them in the queue.
this will prevent jobs from the same batch to be queued in parallel and as a result, hitting ESI queries limit.

also catch TemporaryEsiOutageException when it appends and push back the job in the queue.