Fix error in Seat\Eveapi\Jobs\Assets\Character\Assets
Fixes the error count(): Argument #1 ($value) must be of type Countable|array, stdClass given {"exception":"[object] (TypeError(code: 0): count(): Argument #1 ($value) must be of type Countable|array, stdClass given at /var/www/seat/vendor/softcreatr/jsonpath/src/Filters/SliceFilter.php:17).

When the type of a value passed to the `$overrides` parameter of `AssetMapping::make` is not of type `Closure`, it is treated as a key to be checked on the API response and not a static value. When including data in the asset mapping that doesn't come from the API response, it must be passed as part of a closure to avoid this.